refactor: routeファイル、サイドバーの作成、ディレクトリ構造修正
