+++ frontend/src/components/popUp/create_channel.tsx
@@ -1,4 +1,4 @@
-import { getToken } from "./cookie";
+import { getToken } from "@fetchAPI/cookie";
 
 export interface Channel {
   id: number;
