+++ frontend/src/components/sideNav1/create.tsx
@@ -13,7 +13,6 @@ function CreateWorkspace() {
     let workspaceName = name;
     postWorkspace(workspaceName);
   };
-_
   return (
     <div className="CreateWorkspace">
         <h2>Create Workspace</h2>
