+++ frontend/src/components/sideNav1/workspace_index.tsx
@@ -37,6 +37,7 @@ function WorkspaceIndex() {
       <div>
         {list}
       </div><br></br>
+      
     </div>
     );
 }
