+++ frontend/src/pages/main/login_form.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useCookies } from "react-cookie";
-import { LoginForm } from "@components/user";
+import { LoginForm } from "@src/components/main/user";
 import { Link } from 'react-router-dom';
 import router from "next/router";
