+++ frontend/src/pages/main/signUp_form.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SignUpForm from "src/components/signUp_form";
+import SignUpForm from "@src/components/main/signUp_form";
 export default function SignUp() {
   return (
       <main>
